TimeAndDims does not implement equals/hashcode
