+++ processing/src/main/java/io/druid/query/groupby/GroupByQueryHelper.java
@@ -31,9 +31,8 @@ import io.druid.granularity.QueryGranularity;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.dimension.DimensionSpec;
 import io.druid.segment.incremental.IncrementalIndex;
+import io.druid.segment.incremental.IncrementalIndices;
 import io.druid.segment.incremental.IndexSizeExceededException;
-import io.druid.segment.incremental.OffheapIncrementalIndex;
-import io.druid.segment.incremental.OnheapIncrementalIndex;
 
 import java.nio.ByteBuffer;
 import java.util.List;
@@ -80,36 +79,25 @@ public class GroupByQueryHelper
           }
         }
     );
-    final IncrementalIndex index;
-
+    final boolean useOffheap = query.getContextValue("useOffheap", false);
     final boolean sortResults = query.getContextValue(CTX_KEY_SORT_RESULTS, true);
+    final int maxRowCount = Math.min(
+        query.getContextValue(CTX_KEY_MAX_RESULTS, config.getMaxResults()),
+        config.getMaxResults()
+    );
 
-    if (query.getContextValue("useOffheap", false)) {
-      index = new OffheapIncrementalIndex(
-          // use granularity truncated min timestamp
-          // since incoming truncated timestamps may precede timeStart
-          granTimeStart,
-          gran,
-          aggs.toArray(new AggregatorFactory[aggs.size()]),
-          false,
-          true,
-          sortResults,
-          Math.min(query.getContextValue(CTX_KEY_MAX_RESULTS, config.getMaxResults()), config.getMaxResults()),
-          bufferPool
-      );
-    } else {
-      index = new OnheapIncrementalIndex(
-          // use granularity truncated min timestamp
-          // since incoming truncated timestamps may precede timeStart
-          granTimeStart,
-          gran,
-          aggs.toArray(new AggregatorFactory[aggs.size()]),
-          false,
-          true,
-          sortResults,
-          Math.min(query.getContextValue(CTX_KEY_MAX_RESULTS, config.getMaxResults()), config.getMaxResults())
-      );
-    }
+    final IncrementalIndex index = IncrementalIndices.create(
+        useOffheap ? bufferPool : null,
+        // use granularity truncated min timestamp
+        // since incoming truncated timestamps may precede timeStart
+        granTimeStart,
+        gran,
+        aggs.toArray(new AggregatorFactory[aggs.size()]),
+        false,
+        true,
+        sortResults,
+        maxRowCount
+    );
 
     Accumulator<IncrementalIndex, T> accumulator = new Accumulator<IncrementalIndex, T>()
     {
