+++ processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
@@ -24,7 +24,6 @@ import com.google.common.base.Function;
 import com.google.common.base.Strings;
 import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -38,35 +37,20 @@ import io.druid.data.input.MapBasedRow;
 import io.druid.data.input.Row;
 import io.druid.data.input.impl.DimensionSchema;
 import io.druid.data.input.impl.DimensionsSpec;
-import io.druid.data.input.impl.NewSpatialDimensionSchema;
 import io.druid.data.input.impl.SpatialDimensionSchema;
 import io.druid.granularity.QueryGranularity;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.PostAggregator;
-import io.druid.query.dimension.DimensionSpec;
-import io.druid.query.extraction.ExtractionFn;
-import io.druid.segment.ColumnSelectorFactory;
-import io.druid.segment.DimensionSelector;
-import io.druid.segment.FloatColumnSelector;
-import io.druid.segment.LongColumnSelector;
 import io.druid.segment.Metadata;
-import io.druid.segment.ObjectColumnSelector;
-import io.druid.segment.column.Column;
 import io.druid.segment.column.ColumnCapabilities;
 import io.druid.segment.column.ColumnCapabilitiesImpl;
 import io.druid.segment.column.ValueType;
-import io.druid.segment.data.IndexedInts;
-import io.druid.segment.serde.ComplexMetricExtractor;
-import io.druid.segment.serde.ComplexMetricSerde;
-import io.druid.segment.serde.ComplexMetrics;
 import org.joda.time.DateTime;
 import org.joda.time.Interval;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.GuardedBy;
 import java.io.Closeable;
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
@@ -83,261 +67,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>, Closeable
 {
   private volatile DateTime maxIngestedEventTime;
-
-  // Used to discover ValueType based on the class of values in a row
-  // Also used to convert between the duplicate ValueType enums in DimensionSchema (druid-api) and main druid.
-  private static final Map<Object, ValueType> TYPE_MAP = ImmutableMap.<Object, ValueType>builder()
-      .put(Long.class, ValueType.LONG)
-      .put(Double.class, ValueType.FLOAT)
-      .put(Float.class, ValueType.FLOAT)
-      .put(String.class, ValueType.STRING)
-      .put(DimensionSchema.ValueType.LONG, ValueType.LONG)
-      .put(DimensionSchema.ValueType.FLOAT, ValueType.FLOAT)
-      .put(DimensionSchema.ValueType.STRING, ValueType.STRING)
-      .build();
-
-  private static final Function<Object, String> STRING_TRANSFORMER = new Function<Object, String>()
-  {
-    @Override
-    public String apply(final Object o)
-    {
-      return String.valueOf(o);
-    }
-  };
-
-  private static final Function<Object, Long> LONG_TRANSFORMER = new Function<Object, Long>()
-  {
-    @Override
-    public Long apply(final Object o)
-    {
-      if (o == null) {
-        return null;
-      }
-      if (o instanceof String) {
-        try {
-          return Long.valueOf((String) o);
-        }
-        catch (NumberFormatException nfe) {
-          throw new ParseException(nfe, "Unable to parse value[%s] as long in column: ", o);
-        }
-      }
-      if (o instanceof Number) {
-        return ((Number) o).longValue();
-      }
-      return null;
-    }
-  };
-
-  private static final Function<Object, Float> FLOAT_TRANSFORMER = new Function<Object, Float>()
-  {
-    @Override
-    public Float apply(final Object o)
-    {
-      if (o == null) {
-        return null;
-      }
-      if (o instanceof String) {
-        try {
-          return Float.valueOf((String) o);
-        }
-        catch (NumberFormatException nfe) {
-          throw new ParseException(nfe, "Unable to parse value[%s] as float in column: ", o);
-        }
-      }
-      if (o instanceof Number) {
-        return ((Number) o).floatValue();
-      }
-      return null;
-    }
-  };
-
-  private static final Map<ValueType, Function> VALUE_TRANSFORMS = ImmutableMap.<ValueType, Function>builder()
-      .put(ValueType.LONG, LONG_TRANSFORMER)
-      .put(ValueType.FLOAT, FLOAT_TRANSFORMER)
-      .put(ValueType.STRING, STRING_TRANSFORMER)
-      .build();
-
-  public static ColumnSelectorFactory makeColumnSelectorFactory(
-      final AggregatorFactory agg,
-      final Supplier<InputRow> in,
-      final boolean deserializeComplexMetrics
-  )
-  {
-    return new ColumnSelectorFactory()
-    {
-      @Override
-      public LongColumnSelector makeLongColumnSelector(final String columnName)
-      {
-        if (columnName.equals(Column.TIME_COLUMN_NAME)) {
-          return new LongColumnSelector()
-          {
-            @Override
-            public long get()
-            {
-              return in.get().getTimestampFromEpoch();
-            }
-          };
-        }
-        return new LongColumnSelector()
-        {
-          @Override
-          public long get()
-          {
-            return in.get().getLongMetric(columnName);
-          }
-        };
-      }
-
-      @Override
-      public FloatColumnSelector makeFloatColumnSelector(final String columnName)
-      {
-        return new FloatColumnSelector()
-        {
-          @Override
-          public float get()
-          {
-            return in.get().getFloatMetric(columnName);
-          }
-        };
-      }
-
-      @Override
-      public ObjectColumnSelector makeObjectColumnSelector(final String column)
-      {
-        final String typeName = agg.getTypeName();
-
-        final ObjectColumnSelector<Object> rawColumnSelector = new ObjectColumnSelector<Object>()
-        {
-          @Override
-          public Class classOfObject()
-          {
-            return Object.class;
-          }
-
-          @Override
-          public Object get()
-          {
-            return in.get().getRaw(column);
-          }
-        };
-
-        if (!deserializeComplexMetrics) {
-          return rawColumnSelector;
-        } else {
-          if (typeName.equals("float")) {
-            return rawColumnSelector;
-          }
-
-          final ComplexMetricSerde serde = ComplexMetrics.getSerdeForType(typeName);
-          if (serde == null) {
-            throw new ISE("Don't know how to handle type[%s]", typeName);
-          }
-
-          final ComplexMetricExtractor extractor = serde.getExtractor();
-          return new ObjectColumnSelector()
-          {
-            @Override
-            public Class classOfObject()
-            {
-              return extractor.extractedClass();
-            }
-
-            @Override
-            public Object get()
-            {
-              return extractor.extractValue(in.get(), column);
-            }
-          };
-        }
-      }
-
-      @Override
-      public DimensionSelector makeDimensionSelector(
-          DimensionSpec dimensionSpec
-      )
-      {
-        return dimensionSpec.decorate(makeDimensionSelectorUndecorated(dimensionSpec));
-      }
-
-      private DimensionSelector makeDimensionSelectorUndecorated(
-          DimensionSpec dimensionSpec
-      )
-      {
-        final String dimension = dimensionSpec.getDimension();
-        final ExtractionFn extractionFn = dimensionSpec.getExtractionFn();
-
-        return new DimensionSelector()
-        {
-          @Override
-          public IndexedInts getRow()
-          {
-            final List<String> dimensionValues = in.get().getDimension(dimension);
-            final ArrayList<Integer> vals = Lists.newArrayList();
-            if (dimensionValues != null) {
-              for (int i = 0; i < dimensionValues.size(); ++i) {
-                vals.add(i);
-              }
-            }
-
-            return new IndexedInts()
-            {
-              @Override
-              public int size()
-              {
-                return vals.size();
-              }
-
-              @Override
-              public int get(int index)
-              {
-                return vals.get(index);
-              }
-
-              @Override
-              public Iterator<Integer> iterator()
-              {
-                return vals.iterator();
-              }
-
-              @Override
-              public void close() throws IOException
-              {
-
-              }
-
-              @Override
-              public void fill(int index, int[] toFill)
-              {
-                throw new UnsupportedOperationException("fill not supported");
-              }
-            };
-          }
-
-          @Override
-          public int getValueCardinality()
-          {
-            throw new UnsupportedOperationException("value cardinality is unknown in incremental index");
-          }
-
-          @Override
-          public String lookupName(int id)
-          {
-            final String value = in.get().getDimension(dimension).get(id);
-            return extractionFn == null ? value : extractionFn.apply(value);
-          }
-
-          @Override
-          public int lookupId(String name)
-          {
-            if (extractionFn != null) {
-              throw new UnsupportedOperationException("cannot perform lookup when applying an extraction function");
-            }
-            return in.get().getDimension(dimension).indexOf(name);
-          }
-        };
-      }
-    };
-  }
 
   private final long minTimestamp;
   private final QueryGranularity gran;
@@ -412,7 +141,7 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
 
     for (DimensionSchema dimSchema : dimensionsSpec.getDimensions()) {
       ColumnCapabilitiesImpl capabilities = new ColumnCapabilitiesImpl();
-      ValueType type = TYPE_MAP.get(dimSchema.getValueType());
+      ValueType type = IncrementalIndices.TYPE_MAP.get(dimSchema.getValueType());
       capabilities.setType(type);
       if (dimSchema.getTypeName().equals(DimensionSchema.SPATIAL_TYPE_NAME)) {
         capabilities.setHasSpatialIndexes(true);
@@ -502,27 +231,10 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     return row;
   }
 
-  private ValueType getTypeFromDimVal(Object dimVal)
-  {
-    Object singleVal;
-    if (dimVal instanceof List) {
-      List dimValList = (List) dimVal;
-      singleVal = dimValList.size() == 0 ? null : dimValList.get(0);
-    } else {
-      singleVal = dimVal;
-    }
-
-    if (singleVal == null) {
-      return null;
-    }
-
-    return TYPE_MAP.get(singleVal.getClass());
-  }
-
   private List<Comparable> getRowDimensionAsComparables(InputRow row, String dimension, ValueType type)
   {
     final Object dimVal = row.getRaw(dimension);
-    final Function transformer = VALUE_TRANSFORMS.get(type);
+    final Function transformer = IncrementalIndices.VALUE_TRANSFORMS.get(type);
     final List<Comparable> dimensionValues;
     try {
       if (dimVal == null) {
@@ -1194,6 +906,42 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     public ValueType[] getTypes()
     {
       return types;
+    }
+
+    @Override
+    public boolean equals(Object o)
+    {
+      if (this == o) {
+        return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+        return false;
+      }
+
+      TimeAndDims that = (TimeAndDims) o;
+
+      if (timestamp != that.timestamp) {
+        return false;
+      }
+      if (dims.length != that.dims.length) {
+        return false;
+      }
+      for (int i = 0; i < dims.length; i++) {
+        if (!Arrays.equals(dims[i], that.dims[i])) {
+          return false;
+        }
+      }
+      return true;
+    }
+
+    @Override
+    public int hashCode()
+    {
+      int hash = (int) timestamp;
+      for (int i = 0; i < dims.length; i++) {
+        hash = 31 * hash + Arrays.hashCode(dims[i]);
+      }
+      return hash;
     }
 
     @Override
