+++ processing/src/main/java/io/druid/segment/incremental/IncrementalIndices.java
@@ -1,0 +1,368 @@
+/*
+ * Licensed to Metamarkets Group Inc. (Metamarkets) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Metamarkets licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.druid.segment.incremental;
+
+import com.google.common.base.Function;
+import com.google.common.base.Supplier;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Lists;
+import com.metamx.common.ISE;
+import com.metamx.common.parsers.ParseException;
+import io.druid.collections.StupidPool;
+import io.druid.data.input.InputRow;
+import io.druid.data.input.impl.DimensionSchema;
+import io.druid.granularity.QueryGranularity;
+import io.druid.query.aggregation.AggregatorFactory;
+import io.druid.query.dimension.DimensionSpec;
+import io.druid.query.extraction.ExtractionFn;
+import io.druid.segment.ColumnSelectorFactory;
+import io.druid.segment.DimensionSelector;
+import io.druid.segment.FloatColumnSelector;
+import io.druid.segment.LongColumnSelector;
+import io.druid.segment.ObjectColumnSelector;
+import io.druid.segment.column.Column;
+import io.druid.segment.column.ValueType;
+import io.druid.segment.data.IndexedInts;
+import io.druid.segment.serde.ComplexMetricExtractor;
+import io.druid.segment.serde.ComplexMetricSerde;
+import io.druid.segment.serde.ComplexMetrics;
+
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+
+/**
+ */
+public class IncrementalIndices
+{
+  public static IncrementalIndex create(
+      StupidPool<ByteBuffer> bufferPool,
+      long minTimestamp,
+      QueryGranularity gran,
+      final AggregatorFactory[] metrics,
+      boolean deserializeComplexMetrics,
+      boolean reportParseExceptions,
+      boolean sortFacts,
+      int maxRowCount
+  )
+  {
+    IncrementalIndexSchema indexSchema =
+        new IncrementalIndexSchema.Builder().withMinTimestamp(minTimestamp)
+                                            .withQueryGranularity(gran)
+                                            .withMetrics(metrics)
+                                            .build();
+    if (bufferPool == null) {
+      return new OnheapIncrementalIndex(
+          indexSchema,
+          deserializeComplexMetrics,
+          reportParseExceptions,
+          sortFacts,
+          maxRowCount
+      );
+    }
+    return new OffheapIncrementalIndex(
+        indexSchema,
+        deserializeComplexMetrics,
+        reportParseExceptions,
+        sortFacts,
+        maxRowCount,
+        bufferPool
+    );
+  }
+
+  // Used to discover ValueType based on the class of values in a row
+  // Also used to convert between the duplicate ValueType enums in DimensionSchema (druid-api) and main druid.
+  static final Map<Object, ValueType> TYPE_MAP = ImmutableMap.<Object, ValueType>builder()
+      .put(Long.class, ValueType.LONG)
+      .put(Double.class, ValueType.FLOAT)
+      .put(Float.class, ValueType.FLOAT)
+      .put(String.class, ValueType.STRING)
+      .put(DimensionSchema.ValueType.LONG, ValueType.LONG)
+      .put(DimensionSchema.ValueType.FLOAT, ValueType.FLOAT)
+      .put(DimensionSchema.ValueType.STRING, ValueType.STRING)
+      .build();
+
+  static ValueType getTypeFromDimVal(Object dimVal)
+  {
+    Object singleVal;
+    if (dimVal instanceof List) {
+      List dimValList = (List) dimVal;
+      singleVal = dimValList.size() == 0 ? null : dimValList.get(0);
+    } else {
+      singleVal = dimVal;
+    }
+
+    if (singleVal == null) {
+      return null;
+    }
+
+    return TYPE_MAP.get(singleVal.getClass());
+  }
+
+  static final Function<Object, String> STRING_TRANSFORMER = new Function<Object, String>()
+  {
+    @Override
+    public String apply(final Object o)
+    {
+      return String.valueOf(o);
+    }
+  };
+
+  static final Function<Object, Long> LONG_TRANSFORMER = new Function<Object, Long>()
+  {
+    @Override
+    public Long apply(final Object o)
+    {
+      if (o == null) {
+        return null;
+      }
+      if (o instanceof String) {
+        try {
+          return Long.valueOf((String) o);
+        }
+        catch (NumberFormatException nfe) {
+          throw new ParseException(nfe, "Unable to parse value[%s] as long in column: ", o);
+        }
+      }
+      if (o instanceof Number) {
+        return ((Number) o).longValue();
+      }
+      return null;
+    }
+  };
+
+  static final Function<Object, Float> FLOAT_TRANSFORMER = new Function<Object, Float>()
+  {
+    @Override
+    public Float apply(final Object o)
+    {
+      if (o == null) {
+        return null;
+      }
+      if (o instanceof String) {
+        try {
+          return Float.valueOf((String) o);
+        }
+        catch (NumberFormatException nfe) {
+          throw new ParseException(nfe, "Unable to parse value[%s] as float in column: ", o);
+        }
+      }
+      if (o instanceof Number) {
+        return ((Number) o).floatValue();
+      }
+      return null;
+    }
+  };
+
+  static final Map<ValueType, Function> VALUE_TRANSFORMS =
+      ImmutableMap.<ValueType, Function>builder()
+                  .put(ValueType.LONG, LONG_TRANSFORMER)
+                  .put(ValueType.FLOAT, FLOAT_TRANSFORMER)
+                  .put(
+                      ValueType.STRING,
+                      STRING_TRANSFORMER
+                  )
+                  .build();
+
+  public static ColumnSelectorFactory makeColumnSelectorFactory(
+      final AggregatorFactory agg,
+      final Supplier<InputRow> in,
+      final boolean deserializeComplexMetrics
+  )
+  {
+    return new ColumnSelectorFactory()
+    {
+      @Override
+      public LongColumnSelector makeLongColumnSelector(final String columnName)
+      {
+        if (columnName.equals(Column.TIME_COLUMN_NAME)) {
+          return new LongColumnSelector()
+          {
+            @Override
+            public long get()
+            {
+              return in.get().getTimestampFromEpoch();
+            }
+          };
+        }
+        return new LongColumnSelector()
+        {
+          @Override
+          public long get()
+          {
+            return in.get().getLongMetric(columnName);
+          }
+        };
+      }
+
+      @Override
+      public FloatColumnSelector makeFloatColumnSelector(final String columnName)
+      {
+        return new FloatColumnSelector()
+        {
+          @Override
+          public float get()
+          {
+            return in.get().getFloatMetric(columnName);
+          }
+        };
+      }
+
+      @Override
+      public ObjectColumnSelector makeObjectColumnSelector(final String column)
+      {
+        final String typeName = agg.getTypeName();
+
+        final ObjectColumnSelector<Object> rawColumnSelector = new ObjectColumnSelector<Object>()
+        {
+          @Override
+          public Class classOfObject()
+          {
+            return Object.class;
+          }
+
+          @Override
+          public Object get()
+          {
+            return in.get().getRaw(column);
+          }
+        };
+
+        if (!deserializeComplexMetrics) {
+          return rawColumnSelector;
+        } else {
+          if (typeName.equals("float")) {
+            return rawColumnSelector;
+          }
+
+          final ComplexMetricSerde serde = ComplexMetrics.getSerdeForType(typeName);
+          if (serde == null) {
+            throw new ISE("Don't know how to handle type[%s]", typeName);
+          }
+
+          final ComplexMetricExtractor extractor = serde.getExtractor();
+          return new ObjectColumnSelector()
+          {
+            @Override
+            public Class classOfObject()
+            {
+              return extractor.extractedClass();
+            }
+
+            @Override
+            public Object get()
+            {
+              return extractor.extractValue(in.get(), column);
+            }
+          };
+        }
+      }
+
+      @Override
+      public DimensionSelector makeDimensionSelector(
+          DimensionSpec dimensionSpec
+      )
+      {
+        return dimensionSpec.decorate(makeDimensionSelectorUndecorated(dimensionSpec));
+      }
+
+      private DimensionSelector makeDimensionSelectorUndecorated(
+          DimensionSpec dimensionSpec
+      )
+      {
+        final String dimension = dimensionSpec.getDimension();
+        final ExtractionFn extractionFn = dimensionSpec.getExtractionFn();
+
+        return new DimensionSelector()
+        {
+          @Override
+          public IndexedInts getRow()
+          {
+            final List<String> dimensionValues = in.get().getDimension(dimension);
+            final ArrayList<Integer> vals = Lists.newArrayList();
+            if (dimensionValues != null) {
+              for (int i = 0; i < dimensionValues.size(); ++i) {
+                vals.add(i);
+              }
+            }
+
+            return new IndexedInts()
+            {
+              @Override
+              public int size()
+              {
+                return vals.size();
+              }
+
+              @Override
+              public int get(int index)
+              {
+                return vals.get(index);
+              }
+
+              @Override
+              public Iterator<Integer> iterator()
+              {
+                return vals.iterator();
+              }
+
+              @Override
+              public void close() throws IOException
+              {
+
+              }
+
+              @Override
+              public void fill(int index, int[] toFill)
+              {
+                throw new UnsupportedOperationException("fill not supported");
+              }
+            };
+          }
+
+          @Override
+          public int getValueCardinality()
+          {
+            throw new UnsupportedOperationException("value cardinality is unknown in incremental index");
+          }
+
+          @Override
+          public String lookupName(int id)
+          {
+            final String value = in.get().getDimension(dimension).get(id);
+            return extractionFn == null ? value : extractionFn.apply(value);
+          }
+
+          @Override
+          public int lookupId(String name)
+          {
+            if (extractionFn != null) {
+              throw new UnsupportedOperationException("cannot perform lookup when applying an extraction function");
+            }
+            return in.get().getDimension(dimension).indexOf(name);
+          }
+        };
+      }
+    };
+  }
+}
