+++ processing/src/main/java/io/druid/segment/incremental/OnheapIncrementalIndex.java
@@ -145,7 +145,11 @@ public class OnheapIncrementalIndex extends IncrementalIndex<Aggregator>
     for (AggregatorFactory agg : metrics) {
       selectors.put(
           agg.getName(),
-          new ObjectCachingColumnSelectorFactory(makeColumnSelectorFactory(agg, rowSupplier, deserializeComplexMetrics))
+          new ObjectCachingColumnSelectorFactory(IncrementalIndices.makeColumnSelectorFactory(
+              agg,
+              rowSupplier,
+              deserializeComplexMetrics
+          ))
       );
     }
